Add createButtonX to shared-ui
