+++ services/common/package.json
@@ -54,6 +54,7 @@
     "@emotion/css": "11.1.3",
     "@material-ui/core": "4.11.3",
     "@material-ui/icons": "4.11.2",
+    "@route-codegen/utils": "0.2.0",
     "formik": "2.1.5",
     "graphql": "15.5.0",
     "react": "17.0.1",
